Fixed some stuff
Added a BootstrapException.

Fixed some other dodgy exceptions.

The server now checks if the source IP in the packet matches the IP in
the received origin node object. This reduces unnecessary processing,
future exceptions, and maintains security.

The ConnectReceiver only logs about a new node if the node is actually
a node we’ve never seen before.