See description
Implemented the sending/receiving of the user’s associatedNodes
ArrayList

Changed number of contacts per bucket to 5 - better for testing with
less nodes

Fixed the ‘delay’ that was seen by some nodes whilst receiving messages
which would cause the ACK to timeout on the source. It was caused by
the print line getting the hostname from the node object which was too
expensive. Removed for now.